pyproject.toml
[tool.poetry]
name = "bolagsverket-mcp-server"
version = "0.1.0"
description = "MCP server for the Bolagsverket API"
authors = ["Roo"]

[tool.poetry.dependencies]
python = "^3.9"
mcp = "^0.1.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
